refactor(manifold): extract getInstanceId function to utils file
