+++ packages/manifold/src/utils.ts
@@ -1,4 +1,27 @@
-import { type FilterOperator } from '@rabbitholegg/questdk'
+import {
+  chainIdToViemChain,
+  type FilterOperator,
+} from '@rabbitholegg/questdk-plugin-utils'
+import {
+  ERC1155_CONTRACT,
+  INSTANCEID_ABI,
+  ABI_MINT,
+  ABI_MULTI,
+} from './constants'
+import {
+  createPublicClient,
+  http,
+  type Address,
+  type ContractFunctionReturnType,
+} from 'viem'
+
+export interface ManifoldInput {
+  $abi: typeof ABI_MULTI | typeof ABI_MINT
+  creatorContractAddress: string
+  instanceId?: number | string
+  mintCount?: FilterOperator | undefined
+  mintFor?: string
+}
 
 export const shouldIncludeAbiMint = (
   amount: FilterOperator | undefined,
@@ -13,4 +36,43 @@ export const shouldIncludeAbiMint = (
     return Number(amount) === 1
   }
   return true
+}
+
+const instanceIdCache: Record<string, number | undefined> = {}
+
+export async function getInstanceId(
+  chainId: number,
+  contractAddress: Address,
+  tokenId?: number,
+): Promise<number | undefined> {
+  if (tokenId == null) return undefined
+
+  const cacheKey = `${chainId}-${contractAddress}-${tokenId}`
+  if (instanceIdCache[cacheKey]) {
+    return instanceIdCache[cacheKey]
+  }
+
+  const client = createPublicClient({
+    chain: chainIdToViemChain(chainId),
+    transport: http(),
+  })
+
+  try {
+    const data: ContractFunctionReturnType<typeof INSTANCEID_ABI> =
+      await client.readContract({
+        address: ERC1155_CONTRACT, // ERC-721 should not use tokenId
+        abi: INSTANCEID_ABI,
+        functionName: 'getClaimForToken',
+        args: [contractAddress, tokenId],
+      })
+
+    if (typeof data === 'object' && Array.isArray(data)) {
+      instanceIdCache[cacheKey] = data[0]
+      return data[0]
+    }
+  } catch (e) {
+    instanceIdCache[cacheKey] = 0
+    return 0
+  }
+  return undefined
 }
